moved onnx network operation creation into separate file
